requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.9
pymongo==4.5.0
redis==4.6.0
python-dotenv==1.0.0